updated structure, Cliente fully implemented, JPA.context DONE

+++ F2/src/main/java/isel/sisinf/grp02/repo/IClienteRepository.java
@@ -4,6 +4,6 @@ import isel.sisinf.grp02.JPAObjects.Cliente;
 
 import java.util.Collection;
 
-public interface IClienteRepository extends IRepository<Cliente, Collection<Cliente>, Long>{
+public interface IClienteRepository extends IRepository<Cliente, Collection<Cliente>, Integer>{
 }
